+++ kernel/mem/vm.c
@@ -40,7 +40,7 @@ seginit(void)
     uint32_t base;
   } __attribute__((packed)) gdtr;
   gdtr.limit = sizeof(c->gdt) - 1;
-  gdtr.base = (uint32_t)c->gdt;
+  gdtr.base = (uint32_t)(uintptr_t)c->gdt;
   lgdt(&gdtr);
 #else
   // arch.h version takes 2 parameters
@@ -272,7 +272,7 @@ allocuvm(pde_t *pgdir, uint oldsz, uint newsz)
       return 0;
     }
     memset(mem, 0, PGSIZE);
-    if(mappages(pgdir, (char*)a, PGSIZE, V2P(mem), PTE_W|PTE_U) < 0){
+    if(mappages(pgdir, (char*)(uintptr_t)a, PGSIZE, V2P(mem), PTE_W|PTE_U) < 0){
       cprintf("allocuvm out of memory (2)\n");
       deallocuvm(pgdir, newsz, oldsz);
       kfree(mem);
@@ -297,7 +297,7 @@ deallocuvm(pde_t *pgdir, uint oldsz, uint newsz)
 
   a = PGROUNDUP(newsz);
   for(; a  < oldsz; a += PGSIZE){
-    pte = walkpgdir(pgdir, (char*)a, 0);
+    pte = walkpgdir(pgdir, (char*)(uintptr_t)a, 0);
     if(!pte)
 #ifdef __x86_64__
       a = PGADDR(0, 0, PDX(a) + 1, 0, 0) - PGSIZE;
@@ -497,7 +497,7 @@ exo_unbind_page(exo_cap cap)
   uint pa = e.resource;
 
   for(a = 0; a < p->sz; a += PGSIZE){
-    if((pte = walkpgdir(pgdir, (void*)a, 0)) != 0 && (*pte & PTE_P)){
+    if((pte = walkpgdir(pgdir, (void*)(uintptr_t)a, 0)) != 0 && (*pte & PTE_P)){
       if(PTE_ADDR(*pte) == pa){
         *pte = 0;
         kfree(P2V(pa));
